Do not depend on #method_missing for #respond_to?
After the object is batch loaded, the original method_missing
implementation is not called anymore after a `__replace_with!` call.

When `respond_to?` is called on the loaded object afterwards this will
call `#method_missing` on the loaded object. However, the loaded object
can itself have an implementation of method_missing, in this case only
to determine if it responds to a symbol.

Other than the current implementation, the `#method_missing` could be
listed in `IMPLEMENTED_INSTANCE_METHODS`, so it won't be rerouted to the
lazy loaded object. This however calls `__sync!` again, with the
overhead that entails.

Detected this behaviour first for #4, but that was the wrong patch.